Create reactive display
Modified the mousedown event.

Added an if else loop to check whether the button pressed are numbers or
mathmatical operators.

If they are numbers it runs a concatement function and updates the display.

If it is a mathmatical operator then the operator should be stored (not currently
functional) and sum1 is stored as a number a, then the initial loop runs
again to begin incrementing sum 2 for use as b

//////////////////////// Minor tweaks /////////////////////////

Added a query selector and variable for each button

added variables a => sum1, b => sum2, and operator.  Which will be used
to control which function runs at which time, aswell as be used for
function calls upon the equal press